requirements.txt
asgiref==3.4.1
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
numpy==1.21.4
pandas==1.3.5
psycopg2==2.9.2
PyJWT==2.3.0
python-dateutil==2.8.2
python-magic==0.4.24
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
